pyproject.toml
#
# Project metadata
#

[tool.poetry]
name = "rbot"
version = "0.1.0"
description = ""
authors = ["Julian Da Cunha <j.dacunha@pm.me>"]

#
# Build system
#

[build-system]
requires = ["poetry-core>=1.1.0a6"]
build-backend = "poetry.core.masonry.api"

#
# Entrypoints
#

[tool.poetry.scripts]
rbot = "rbot.main:main"

#
# Project Dependencies
#

[tool.poetry.dependencies]
python = "^3.8"
pyaml = "^20.4.0"
rich = "^10.6.0"
PyNaCl = "^1.4.0"
"discord.py" = "^1.7.3"
PyYAML = "^6.0"
pydantic = "^1.8.2"

#
# Dev Dependencies
#

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
hypothesis = "^6.31.3"

[tool.poetry.group.quality]
optional = true

[tool.poetry.group.quality.dependencies]
flake8 = "^4.0.1"
pyproject-flake8 = "^0.0.1-alpha.2"
flake8-simplify = "^0.14.2"
flake8-return = "^1.1.3"
flake8-pytest-style = "^1.5.1"
flake8-comprehensions = "^3.7.0"
flake8-bugbear = "^21.11.29"
flake8-builtins = "^1.5.3"
flake8-commas = "^2.1.0"
flake8-debugger = "^4.0.0"
flake8-eradicate = "^1.2.0"
pep8-naming = "^0.12.1"
mypy = "^0.910"
pydocstyle = "^6.1.1"
types-PyYAML = "^6.0.1"

[tool.poetry.group.format]
optional = true

[tool.poetry.group.format.dependencies]
isort = "^5.10.1"
black = "^21.12-beta.0"

[tool.poetry.group.security]
optional = true

[tool.poetry.group.security.dependencies]
bandit = "^1.7.1"
safety = "^1.10.3"

[tool.poetry.group.changelog]
optional = true

[tool.poetry.group.changelog.dependencies]
git-changelog = "^0.5.0"

#
# Tools configurations
#

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
target-version = ["py310"]
line-length = 120

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.flake8]
max-line-length = 120

[tool.pydocstyle]
inherit = false
convention = "google"
add-ignore = "D100,D107,D212"
match = "^(?:(?!__)(?!main).)+py$"
#match-dir = "[^a_dir_to_exclude].*"

[tool.pytest.ini_options]
addopts = "--cov-report=xml:coverage.xml --cov-report=term --durations=0 --cov=app"
